requirements.txt
asgiref==3.7.2
Django==5.0.1
django-jasmine==0.5.1
djangorestframework==3.14.0
drf-yasg==1.21.7
hashids==1.3.1
inflection==0.5.1
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
